- Updated LANDMark version - TreeOrdination can now take advantage of LANDMark's proximity measures (both using terminal nodes as features or all nodes in the decision path as features) - Data is cast into the np.float32 dtype for the CLRClosureTransformer

+++ TreeOrdination/transformers_treeord.py
@@ -16,18 +16,18 @@ class CLRClosureTransformer(BaseEstimator, TransformerMixin):
     def fit_transform(self, X, y=None, **kwargs):
 
         if self.do_clr:
-            return clr(multiplicative_replacement(closure(X), delta=self.delta))
+            return clr(multiplicative_replacement(closure(X), delta=self.delta)).astype(np.float32)
 
         else:
-            return closure(X)
+            return closure(X).astype(np.float32)
 
     def transform(self, X, y=None, **kwargs):
 
         if self.do_clr:
-            return clr(multiplicative_replacement(closure(X), delta=self.delta))
+            return clr(multiplicative_replacement(closure(X), delta=self.delta)).astype(np.float32)
 
         else:
-            return closure(X)
+            return closure(X).astype(np.float32)
 
 
 class ResampleRandomizeTransform(BaseEstimator, TransformerMixin):
